+++ Cargo.toml
@@ -48,17 +48,15 @@ schemars = { version = "0.8", features = ["chrono", "url", "preserve_order"] }
 
 # Async runtime and utilities
 tokio = { version = "1", features = ["rt"] }
-tokio-stream = "0.1"
 tokio-util = "0.7"
 futures = "0.3"
 futures-util = "0.3"
-async-trait = "0.1"
 
 # Network and API
 axum = { version = "0.8.4", features = ["macros", "ws"] }
-tower = { version = "0.4.13", features = ["util"] }
+tower = { version = "0.5.2", features = ["util"] }
 http-body-util = "0.1.3"
-tokio-tungstenite = "0.22.0"
+tokio-tungstenite = "0.26.2"
 
 # Logging and tracing
 tracing = "0.1"
@@ -77,5 +75,5 @@ bon = "3.6.3"
 regex = "1"
 url = { version = "2", features = ["serde"] }
 chrono = { version = "0.4", features = ["serde"] }
-uuid = { version = "1.16.0", features = ["v4", "serde"] }
+uuid = { version = "1.17.0", features = ["v4", "serde"] }
 rand = "0.8"
